requirements.txt
mkdocs
mkdocs-material
fontawesome_markdown
markdown-include
Pygments>=2.2
pymdown-extensions
